[Refactor] 블로그 테이블 분리
